history_service/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
requests
python-dotenv
pillow
python-consul
pika
